removed find of string component
