+++ packages/api-explorer/src/components/Response/style.js
@@ -1,0 +1,15 @@
+import colors from '../../colors'
+
+const notJsonStyle = {
+    fontSize: 12,
+    fontFamily: 'Monaco, "Lucida Console", monospace',
+    border: 0,
+    background: 'transparent',
+    padding: '0 30px',
+    overflow: 'visible',
+    whiteSpace: 'pre-wrap',
+    wordBreak: 'break-all',
+    color: colors.notJson
+}
+
+module.exports = notJsonStyle
